Backend/requirements.txt


## Note: Python Version 3.11.1 Required
## Note: Install ffmpeg with homwbrew 



pydub
openai
python-dotenv
yt-dlp
groq
moviepy
